requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pycryptodome==3.6.3
requests==2.19.1
tornado==5.0.2
Tornado-MySQL==0.5.1
urllib3==1.23
